backend/tools/youtube/requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
